split huge cpp file to smaller ones
